requirements.txt
flask-migrate
python-barcode
Pillow
flask-wtf
mercadopago
flask-mail
gunicorn





